refactor: remove defaultAppearance constant and update imports across components
