--- src/constants/defaultAppearance.ts
@@ -1,11 +1,0 @@
-import { IAppearance } from '../types';
-
-export const defaultAppearance: IAppearance = {
-  theme: 'light',
-  background: '#ffffff',
-  accent: '#0C1083',
-  textColor: '#000000',
-  font: 'Inter',
-  cornerRadius: 'full',
-  cover: '',
-};
