+++ src/constants/index.ts
@@ -1,4 +1,5 @@
 import { ModalHeights, Routes } from '../types';
+import { IAppearance } from '../types';
 
 export const MODAL_HEIGHTS: ModalHeights = {
   [Routes.PROFILE]: 469,
@@ -8,4 +9,14 @@ export const MODAL_HEIGHTS: ModalHeights = {
   [Routes.ONBOARDING]: 399,
   [Routes.SEND]: 356,
   [Routes.ACTIVITY]: 356,
+};
+
+export const defaultAppearance: IAppearance = {
+  theme: 'light',
+  background: '#ffffff',
+  accent: '#0C1083',
+  textColor: '#000000',
+  font: 'Inter',
+  cornerRadius: 'full',
+  cover: '',
 };
