+++ src/index.ts
@@ -2,4 +2,4 @@ import './tailwind.css';
 
 export { useBlux } from './hooks/useBlux';
 export { BluxProvider } from './context/provider';
-export { defaultAppearance } from './constants/defaultAppearance';
+export { defaultAppearance } from './constants';
